Refactor app.js and homePage.js
